search: introduce Debug field and feature flag (#43782)
This feature flag allows us to set DebugScore on zoekt.SearchOptions.
This will help us understand ranking in the frontend. Right now we only
pass Debug back on FileMatch (and not on individual chunk matches), but
this will already greatly help us. In future PRs we can work on showing
this information in the frontend and allowing this to be set outside of
feature flags.

This is an abuse of the Feature flag struct. However, it is very
convenient for now and is a relatively clean use of it.

I'll first follow up by getting src-cli to print this debug info.

Test Plan: set the feature flag then curl streaming search endpoint and
see debug messages.