Avoid session deletion when sessionstore host can't be found (#43856)
The previous check was specific to "connection refused" errors. But it's
possible that the connection fails due to other network errors too, such
as "dial tcp: lookup horsegraph: no such host".

In those cases we also don't want to delete the session.

So what I did here is to make the check less specific and check for
`net.OpError`. That catches both errors and more.