featureflag: add per request overrides
This is inspired by our ?trace=1 query parameter, except for
featureflags in a request. I want to quickly toggle a feature flag to
compare before and after.

I had behaviour like this at FB, except it was a more general topic
called cheat codes which worked across many systems. We should probably
invest in that as well, but for now this will make it quick to toggle
boolean values in our codebase.