Remove useless status messages tests (#43941)
The tests are useless because they don't test anything. If a repository
is in `cloneStatus` but not in `repos` in the test then the clones
status write will be skipped.

And since that kind of setup doesn't have an effect the tests become
duplicates from other tests and are thus useless.