search: display debug messages in search results
When the search-debug feature flag is enabled, we send down an extra
"debug" field. This contains information on how ranks are computed. This
is extremely useful to understand why we ranked results the way we did.

Right now the presentation is super ugly. But I don't know any better
and this is only for consumption of search engineers.

Test Plan: enabled "search-debug" feature flag on local server and
observed debug information on results pages.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>