Update index.md
The checkbox labeled "Copy the master branch only" was automatically checked when I was working through these steps. This prevented the repo version tags from being copied over and created an error in a later step when trying to create a branch from a specific version. This is to make a note in the guide to prevent this for future users.