include username/email in in-product feedback submissions
The "Feedback" nav item and other feedback forms now submit the current user's username and email address so we can act on the feedback more effectively and follow up for more details or to confirm that an issue is fixed.

For authenticated users, it clearly indicates that (1) their username and email address are being sent along with their feedback and (2) that it's being sent *to Sourcegraph* (and not just to their internal admin).

For anonymous users, it asks them to include contact information in their message so we can reply to them.