don't show the NPS survey on Sourcegraph.com
This change only affects Sourcegraph.com, not customer instances (where the NPS survey feature will continue as-is).

On Sourcegraph.com, we will rely instead on the (now improved) global nav feedback form for qualitative feedback, and on telemetry for quantitative feedback. It is confusing to have 2 different feedback forms.

On customer instances, the NPS survey is still valuable because more feedback is needed in specific time windows, and telemetry is limited.