web: use Headers type for graphql getHeaders
This change is to make it possible to use "Headers.append" so we can set
multiple values for a single header. According to the type documentation
and the typescript compiler this should work.

However, in practice this breaks the E2E tests. Every graphql seems to
fail with 401 Unauthorized. This makes me suspect that somehow doing
"...window?.context?.xhrHeaders" into the construct of Headers is
working incorrectly.

For more context, this is what I want to add in:

  for (const feature of searchParameters.getAll('feat')) {
    headers.append('X-Sourcegraph-Override-Feature', feature)
  }

Alternatively I could probably not use Header and just encode the values
for the header myself. But I'm interested why this is not working.

Test Plan: main dry run which passes (to exercise E2E tests)