Enable use of executor secrets for SSBC (#43940)
This PR makes use of the generic executor secrets infrastructure in SSBC! We will now read the secret values for resolving workspaces and when spawning an executor job, to make sure that everything is set as required and that the cache keys are properly calculated so that a secret value change would invalidate the cache.

For this to have any effect, src-cli will need to be bumped to a new version. It will not break existing executions though so this is safe to merge.