client: conditionally set X-Sourcegraph-Should-Trace (#44061)
Currently we always set X-Sourcegraph-Should-Trace. However, based on
how the backend works, we should only be setting this if the trace
parameter has been specified by the user.

Note: in other places we set this header we only set when it is
specified. So only graphql had this behaviour.

Note: I pulled out the parameter construction so it can be reused. This
will be used in an upcoming commit for "feat".

Test Plan: main dry run to ensure we don't break E2E tests