search: enable large count: queries with ranking
This will degrade the ranking experience in general, but this gives us
similiar behaviour to before were we look for more results if there is a
higher count.

This is different to our non-ranking behaviour in one key way. We are
not using generic scaling factor, instead we just increase in line with
limit once the amount of work we ask zoekt to do is larger than the
default.

Test Plan: CI