Fix FilteredConnection pagination race condition (#44091)
When any search value changes (ie. ?visible=20), this would trigger and cause a refetch. This would refetch before the visible param is updated, so right after the request for more data finishes, this would trigger another fetch for the initial data.
This fixes that by making sure that it only triggers when actually a filter value changed.