improve in-product feedback collection (#43984)
* remove SourcegraphExtensionFeedback

The extension registry is deprecated, and it's not worth updating this usage of the FeedbackPrompt for something that will affect almost zero users.

* include username/email in in-product feedback submissions

The "Feedback" nav item and other feedback forms now submit the current user's username and email address so we can act on the feedback more effectively and follow up for more details or to confirm that an issue is fixed.

For authenticated users, it clearly indicates that (1) their username and email address are being sent along with their feedback and (2) that it's being sent *to Sourcegraph* (and not just to their internal admin).

For anonymous users, it asks them to include contact information in their message so we can reply to them.

* openByDefault=true in FeedbackPrompt storybook

* show product research message only to users

Also lets you toggle productResearchEnabled in storybooks so you can test this variant more easily.

* remove rating emojis (sad/happy) from feedback prompt

The rating emojis in the feedback prompt (very sad, sad, happy, very happy) added friction to the process of submitting feedback. Their meaning was not clear, and we were not using the data.

Now, the feedback prompt has just a single free-form textarea.

* don't show the NPS survey on Sourcegraph.com

This change only affects Sourcegraph.com, not customer instances (where the NPS survey feature will continue as-is).

On Sourcegraph.com, we will rely instead on the (now improved) global nav feedback form for qualitative feedback, and on telemetry for quantitative feedback. It is confusing to have 2 different feedback forms.

On customer instances, the NPS survey is still valuable because more feedback is needed in specific time windows, and telemetry is limited.

* show feedback in global navbar for all users (not just authed)