client: always append preserved query parameters
This simplifies the implementation for preserving the URL parameters
across search requests. Initially I adjusted the buildSearchURLQuery
implementation, but this ended up being a bit ugly and would of only
been used by one call site.

So instead I adjusted how we build the list of preserved parameters and
just appended those. I believe this implementation is more robust and
easier to understand. Additionally we won't undo the escaping we
previously did if a preserved parameter is present.

Test Plan: CI