Revert "prometheus: increase scrape and evaluation intervals to 1m" (#44168)
Reverts #41872

This is technically never respected because k8s has always had an override and a similar override was recently added for docker-compose. Additionally, it's better for metrics that sample on 1m if we have a 30s interval (see PR description: #44166).

This also adds links for reference since the relationship can be confusing.