web: link to codehost when file is an LFS Pointer (#43686)
We extend the GraphQL schema to optionally have a field LFS.
This field is non-null when the content of the git blob is a LFS
pointer. When it is we show a page linking to the code host. This
was a request from a customer.

The actual git lfs project uses similiar heuristics, except it does 
expect the full document to be this KV style. Other projects such
as gitlab just use regexes and file size like we do.

Test Plan: manually visited in dev server LFS pointer files.

Co-authored-by: Juliana Peña <me@julip.co>