searcher: record prometheus metrics for hybrid search
We currently have detailed logs for hybrid search, but as we get
confident in the feature and scale out we need aggregate statistics to
guide us on how it is behaving.

These metrics will help us understand how often hybrid search is
successful and when it isn't, why.

Test Plan: CI