gitserverproxy: Handle panic
When the remote resp.Body cannot be read, the http proxy panics with `ErrAbortHandler`. That is, because at that point in time, the HTTP headers have already been sent and there is no other way out. This will then cause a connection reset error.
Since this error is not "as bad", we should recover it and log something useful instead.