searcher: remove PathPatternsAreRegExps
We have hardcoded this value to true in the client for a very long time.
This PR removes the field and updates conditionals as if the field was
true. Additionally test cases based on glob fields are updated to pass
in the equivalent regex.

A follow-up PR will remove glob support from our path matcher package.

Test Plan: go test

plz-review-url: https://plz.review/review/14688