monitoring: searcher group for hybrid search (#44355)
We export two specific metrics for hybrid search and add a hidden group
showing the metrics. We included a detailed interpretation of the
results aswell as emphasize the experimental nature of the feature.

The plan is once the feature is turned on by default it won't be hidden
by default.

Test Plan: ran the local monitoring stack, but connected to prod. Also
will check the output of markdown in the github renderer.