PermsSyncer: report exact state of each provider in each job (#44257)
Right now there are three ways to try and determine if an authz/authorization provider is happy, each with caveats:

- through metrics and logs exported by each individual provider implementation. This is very inconsistent, up to each implementation, and difficult to programmatically assess
- through `PermsSyncer` job results, which only reports the overall state of a sync job run in a single `error`. It sometimes doesn't report any error from the provider at all, swallowing the provider's error with special handling
- through `PermsSyncer` metrics and logs. This gets us ~almost there, but is difficult to programmatically assess, requires infrastructure access, and requires knowledge of how `PermsSyncer` is implemented.

This makes assessing the health of authz difficult when:

- attempting to validate a particular authz provider is behaving correctly - e.g. when restoring from a backup, we want to be able to select a particular provider to ensure permissions sync jobs are correctly involving the provider and succeeding
- hypothetically, when multiple authz providers are configured

This change propagates the "real" state of each individual provider from the various `runSync` implementations, so that we can see the state of each provider involved in a permissions sync job. As a simple proof-of-concept, I've added a log field that lists the providers that failed when a job fails, and added some assertions to a few tests to demonstrate what the output looks like.