search: Add user whitelist option for search sanitization (#44244)
* allow site admins to name an org that functions as an allowlist for search sanitization
Make the search sanitization experimental feature an object in site schema and provide the field `orgName` which site admins can use to specify an org whose members should not have their search results sanitized. During the Plan step, check if the current user's searches should be sanitized.

* reword description in schema to be clear that sanitization is always off for site admins

* inject compiled sanitization patterns into SearchInputs directly
Rather than calling conf.Get() during both the Plan and Execute steps, only call it once during the Plan step and compile the sanitization patterns to regexps at the same time, returning the slice of compiled regexps and setting them as a field on SearchInputs. Check the length of that field during job construction; if it's nil or empty then sanitizeJob won't run.

* update schema comment