sg: ci build: forcefully set a branch for external commits
Using `sg ci build -c $commit` when you're on the main branch, like when
you're requesting a build for a PR from an external contributor, would
previously end up on the `main` branch in Buildkite, even though it's
actually not on the `main` branch in the repository.

This is quite scary for the random bystander and in the end, very
confusing and wrong.

This change merely hotfixes the problem and the `sg ci build` command
should be rewritten as we also discovered other issues such as
`sg ci build main-dry-run -c $commit` will actually push the current
commit and not $commit on `main-dry-run/main`, which also quite
confusing.