gqltest: Ensure fusion test is actually cloning depot (#44539)
The previous tests we not testing cloning via p4-fusion, the following was happening:

git p4 test runs, completes and deletes external service which then soft
deletes repo. p4-fusion can't find repo to delete since it's name has changed
so does nothing. Repo is still on disk so we end up fetching rather than
cloning. Even if that fails, we still find the repo on disk and mark the test
as passed.

To fix this we now:

* Delete the repo on disk before deleting the service. (Confirmed locally that the repo is removed on disk)
* Clone a different repo in each test
The tests were also refactored to make things a little more explicit.