searcher: handle context errors in hybrid
We have a fair number of errors in hybrid search, but they all seem to
be context canceled. This happens because in the frontend we spin up
many searcher requests (per repo) and then cancel the RPC once we hit a
limit. Additionally searcher itself will cancel after we hit a limit.

Test Plan: CI and then will monitor graphs in production. I expect
zoekt-search-error to go to zero always.