syncx: import code from sync.Once* proposal
This is the code from https://go.dev/cl/451356 renamed into an internal
syncx package. This allows us to use the feature before it lands in the
go stdlib.

As an example I've updated some call sites under cmd/gitserver to
replace usages of sync.Once.

Note: not all sync.Once usages can be replaced, but a majority can be in
our codebase. The exception to this is a lot of our graphql resolvers
use the first context.Context passed in, so wouldn't work with these
utility functions.

Test Plan: go test