monitoring/zoekt: do not manipulate targets for multiple metrics in panel (#44621)
Custom panel options are currently used to hijack the targets to enable multiple queries on this panel - this is not something we support. Especially with the recent changes to leverage Query to create allowlist metrics we collect on Cloud, doing this might mean certain metrics do not get collected in Cloud. It also seems to cause funky interactions with some manipulations we do to queries.

Instead, panels like this should use a query on __name__ - the legend isn't quite as nice, but a follow-up PR could add a label_replace to format it better.

Doing this adds a metric to our allowlist that was previously omitted and not collected in Cloud (in the metrics list we include this entry as index_num_assigned|index_num_indexed|index_queue_cap, which is fine because we use go run ./dev/sg monitoring metrics -format regexp for the allowlist which still ends up with a valid regexp)