Prevent double-fetching user before rendering UserArea
There's a race condition on slower networks where authenticatedUser would be null initially and the user area user request has to run, then the siteAdmin flag changes to true, and then it has to run again before the UI is properly rendered. This makes it a bit faster. Also added a missing loader that caused a small blip that felt weird.