batches: Create page replaces history
Instead of pushing to the history, we should replace it with the spec stage and then replace it to navigate to the edit step. This feels more natural to me than going back in history and coming back to the create page, since the create page is really just "step 1" of the process and morphs into that step 1.