searcher: always handle request cancellation in hybrid metrics (#44818)
When looking at dashboards we have a few errors across our different
places we can fail. However, in everycase the logs seem to correlate
with request cancellation (as well as a high volume of requests).

As such we ensure we always handle these conditions when recording
metrics.

Test Plan: go test as well as monitoring grafana in prod.