Fix workspace resolution infinite loading (#44833)
For a new batch change, and when the current batch spec has been applied before, the workspace resolution would be trapped with an infinite loader.
The initial empty spec has no workspace resolution. When submitted, we start fetching the status of the batch spec known prior saving. For ReplaceBatchSpecInput, that is not a problem, but when we go the CreateBatchSpecFromRaw path, we need to fetch the state of that NEW batch specs workspace resolution.