search: make `repo:has(key:)` valid syntax (#44868)
make `repo:has(key:)` valid syntax
If no value is specified in the predicate argument then it will filter to repos with a matching key and an empty string as the value. This is distinct from `repo:has.tag(key)` which will filter to repos with a matching key and a null value.