monitoring: add stub implementation for multi-instance dashboard (#44748)
Lays some groundwork for generating a new dashboard of panels that aggregate metrics per-instance across multiple instances. The result will be that engineers can opt-in their panel to be included on this multi-instance dashboard with `(Observable).MultiInstance`, upon which we:

- add `project_id` (via parameterized groupings) to all aggregation expressions on the observable panel
- add `- {{project_id}}` (via parameterized groupings) to all label templates on the observable panel
- add the observable panel to a row representing the dashboard in the multi-instance dashboard

The opt-in is mainly because many dashboards do not work well when aggregated per-Cloud-instance. For example, a per-gitserver-instance will have thousands of time series when aggregated per-Cloud-instance, and similarly for other high-dimensionality. This allows us to curate a set of dashboards that give useful cross-instance signals. Eventually, we want to add template variables to allow selecting by instance.

This feature is very much a work-in-progress, but merging to iterate in main.