github perm syncing: remove useless additional request when listing collabs
I noticed that when looking at "Outbound requests" in the site-admin
panel that we send 2 requests to list collaborators, even if that
repository is a private repository of mine with a single collaborator.

Since we list 100 collaborators per page - why would we need to send two
requests to list 1 collaborator?

So I dug in and found that we always do a second request if we have >0
collaborators returned, which doesn't make a lot of sense.

This fix is just something I hacked together, there might (should?) be a
better way to determine whether there is a next page.