CodeMirror blob: implement "navigate to any line on click"
This commits makes it possible to promote any line from the ref panel
by clicking anywhere on that line in the preview pane. This feature is
supported by the old blob view, this commit just closes the feature gap.

This commit implements this functionality by wrapping all lines with an
anchor link. This is slightly different from the old blob view where
every line registers an ad-hoc click handler.