remove unnecessary warning for non-HTTPS externalURL
When the externalURL is `http://` (non-HTTPS), a warning was shown: "Your connection is not private. We recommend ...". This warning was obtrusive and annoying when you are trying Sourcegraph on localhost, haven't set up HTTPS, or are accessing Sourcegraph via a separate HTTPS proxy but haven't set up externalURL to use the HTTPS URL.

I believe this warning is security theater. Site admins should know when they need to set up HTTPS, their browser will warn them, and an https externalURL is not a guarantee of security anyway. For reference, GitLab does not display a warning when using HTTP: https://docs.gitlab.com/omnibus/settings/ssl/ ("By default, HTTPS is not enabled.").