call env.Lock() outside of shared main
Shared main functions allow "programs" to be invoked from different entrypoints. We use this to have separate OSS and non-OSS programs for frontend, repo-updater, etc.

This updates all programs except for worker, which will take some more refactoring to make this change (because it calls `loadConfigs` in the shared main).

I would like to be able to invoke multiple shared main functions in the same program, which requires moving global/singleton things such as `env.Lock()` (which may be called at most once) outside of the shared main function.