QQQ: remove oobmigration from frontend
I think this is safe, but...

Why is oobmigration run both in the frontend and in the worker service? Is this a defensive mechanism in case the worker is not deployed, or is it a legacy holdover from before the worker existed?

(Asked question at https://docs.google.com/document/d/1SHffVw8NrLCmWMXZ_vCFchbiuZUDGAR8vYCoXyJITbE/edit#bookmark=id.8qb9dzx4limx.)