allow duplicate env.Get registrations
In the upcoming single-program mode, multiple services in the same process will call the same env.Get lines. This is safe as long as they register the same default value and description for a given env var name. The alternative would be to ensure that env.Get is only called a single time for each env var name, which would require a lot of pointless "is initialized?" checks and lead to more complexity than this simple change.

Also adds tests for env.Get.