do not panic on duplicate metrics collector registrations
Some metrics are used by multiple services; in particular, worker and precise-code-intel-worker share a lot of metrics.

In the upcoming single-program distribution method, both of these services run in the same Go process. This means that their metrics registrations conflict and result in a panic. This change makes it so it does not panic and the metrics registrations are reused by both.