allow having an empty list of zoekt replicas (means just use unindexed)
If the env var INDEXED_SEARCH_SERVERS is the empty string, then indexed search (Zoekt) will not be used. (If it is merely unset, then it will be used, so the default behavior is unchanged.)

This is useful when running Sourcegraph in environments where indexed search is not needed or desirable. (In particular, the single-program experimental distribution method, but this fix seems useful even on its own.)