remove external URL inference from nginx/http listen addresses (#45118)
* fix capitalization of defaultExternalURL

* remove external URL inference from nginx/http listen addresses

If externalURL was not set in site config, the externalURL would be inferred from the listen address. This is undesirable for several reasons:

- It is a rarely used (because externalURL is almost always set) and untested code path.
- It is not guaranteed to be correct, and an incorrect value here would cause confusion.
- If the externalURL is not set, it's still possible to access the web app and change the site config to fix it.
- The inference is only possible from frontend, not repo-updater, so a default of http://example.com would be use in repo-updater. Different behavior here can lead to a bug.

If an instance relied on this behavior, then the site admin will need to add externalURL to site config. It is extremely unlikely that a real, production instance relied on this behavior because only `http://` URLs would be inferred, and it's unlikely that a real production instance would use `http://`.