search: record zoekt backend being down as a crash (#45067)
This will allow us to track in our observability how often we do a
search which doesn't run on all instances. Additionally this will be
used for a "strict" mode search which fails if any Zoekts are down. This
is intended to be used by code insights.

Test Plan: Updated unit tests