remove old, rarely used, and broken search.index.enabled:false site config (#45138)
The `search.index.enabled` site config setting defaults to true for all deployment types (and has for ~2 years). It was possible to set `"search.index.enabled": false` in site config.

But this was broken; searches with `index:no` would show no results, and normal searches would show an error about index search being disabled. This setting made the instance completely unusable, so it is highly unlikely any instances were actually using it.

The need for this setting has also gone down dramatically with many more years of experience running indexed search at scale on many instances. If we need to bring back the ability to disable indexed search, we should reimplement it (because many code paths did not properly respect this setting).