Implement noCache flag on execute endpoint (#44838)
This PR implements (finally) the no cache flag on the execute endpoint. This means that we can finally allow not refreshing the workspace preview when the checkbox is toggled. For long-running workspace resolutions (like our scaletesting) this is a gamechanger to me (maybe just me who almost always runs without cache because I test execution a lot, but hopefully also someone else 😬).

This also changes a bit how we handle this flag as well:
We don't skip cache checking when no_cache is set anymore, so that the UI always reflects the cache status and you can simply say "don't use it" - this feels more expected to me than some checkbox that lives outside the spec that would modify the workspaces list.
So now the only thing that the workspace resolution does is setting the no_cache flag on the batch spec.
When the batch spec is executed, we check that field, or optionally overwrite it, if set in the mutation arguments.
The field on the createFromRaw,replaceInput mutations is now mostly there to forward that flag to the execute step once we have implemented autoApply and is not used from the UI for now.