sg: use same PG envvar copy for code insights
Directly constructing a DSN is incorrect, since the user may have just
specified a DSN. Instead we need to copy over from the same format that
the user specified for connecting to frontend. This is the same approach
use by codeintel.

Test Plan: sg start enterprise-codeinsights works on my machine (did not
before).