faster function-incomplete local dev builds without cgo
The squirrel package requires cgo, which significantly increases the build (linking) time of the single-program build (from ~3.8s to ~8s). During local development, omitting the local code intel functionality that requires cgo speeds up the build significantly. For production builds, cgo is still needed.

To use this, just run `CGO_ENABLED=0`. Before this commit, that produces a compile error. After this commit, that compiles, but search-based code intel will show an error in the hovers. As such, and because we generally want dev to not diverge from prod builds, this will remain an undocumented dev setting until further notice.