searcher: introduce retry reason for hybrid search (#45096)
We now record in our logs and observability the reason for retrying.
Normally this should be due to the underlying index changing. But there
are other reasons, and as we introduce more it would be great to
differentiate between them.

We rename the metric to make this clear. This feature hasn't launched
yet so safe to rename.

Test Plan: go test