Token selection: implement document highlights (#45205)
* Token selection: implement document highlights

Previously, we used the same document highlight implementation as
the old blob view, which triggered document higlights on every mousemove
event. This commit changes the behavior to trigger document highlights
on token selection mirroring the behavior of IntelliJ and VS Code. One
benefit of this approach is that it's compatible with keyboard
navigation because we trigger higlights based on selection instead of
mouse movements. Personally, I also like that I can now select a token
and scroll through the file to see the matching occurrences (previously,
moving the mouse would make the highlights disappear or change.

* Empty commit to trigger ci

* Use more appropriate function name

Co-authored-by: Taras Yemets <yemets.taras@gmail.com>

* Use better function name

* ESlint

Co-authored-by: Taras Yemets <yemets.taras@gmail.com>