split apart metrics used for precise-code-intel-worker vs. worker
The `newOperations` func was called by both the precise-code-intel-worker and the worker. The `uploadProcessor` and `uploadSizeGuage` (sic) were used by the former but not the latter. This refactor makes it so that each service only instantiates the actual metrics it truly uses and needs.

(This is also useful for the experimental single-program distribution, where we want both services to run in the same process, but it's a plausible refactor on its own as well.)