Token selection: add back "go to definition" button (#45294)
We received internal feedback asking what happened to the "Go to
definition" button. This commits adds this button back to reduce
disruption when rolling out token selection. The new button is no longer
a link like it was before because links are not a suitable abstraction
for "go to definition", as documented in
https://docs.google.com/document/d/1cXZEXfui1ylXBfNoB5iK7sUYhbNm51Y7wQMW36RScTA/edit#

In addition to the "You are at the definition", which the old navigation
supported, this new implementation also displays "No definition found"
when there are guaranteed no results.