database: UsersListOptions.Query searches for ID
A common pain point I have is finding the user for an ID that I come
across in logs or honeycomb. This addition makes it possible to search
for an ID or GraphQL ID in the site admin user panel and if it matches a
user that user will be returned.

Test Plan: added a unit test for List