Fix blame component when class contains plus (#45350)
* Fix blame component when class contains plus

Okay so weird bug here: The sass classname has a suffix that is the file hash, base64 encoded and then the first 5 characters of that.
The base64 charset includes the `+` character, and that character has a special meaning in `querySelector`. This fixes it by using it as the literal classname.