security_event_logs: set AnonymousUserID to internal for internal actors (#45309)
Some instances are seeing security log events failing to insert when both UserID and AnonymousUserID are unset. To work around this, we set internal as the AnonymousUserID for all events that are under an internal actor and don't have an associated user, to avoid violating the constraint. Note that a similar workaround already exists at callsites.