fix creating logger before log init
PR #45369 introduced an issue where new loggers were being created in a
static context. This is a somewhat hacky fix to ensure that loggers are
not created until after initialization. Ideally, the gitservers map is
created with an explicit logger after initialization, but I'm just
aiming to fix sg start right now.