map host.docker.internal in executors to the Docker host
This makes it possible to use executors when externalURL is http://localhost:3080 or similar (which is the case for local installations). In that case, executors.frontendURL needs to be set to http://host.docker.internal:3080, because localhost in the executor container would resolve to the container's localhost.

While the `--add-host=HOST:host-gateway` option is (strangely) not documented in Docker docs, it is supported on Docker for Windows, macOS, and Linux.

Also rewrites the site config description for executor.frontendURL to be a bit clearer.