database: ReposListOptions.Query searches for ID (#45354)
A common pain point I have is finding the repository for an ID that I
come across in logs or honeycomb. This addition makes it possible to
search for an ID or GraphQL ID in the site admin repository page and if
it matches a repository that repository will be returned.

Note: ReposListOptions.Query is only used for repository search in the
admin panel. Normal search uses the include/exclude fields.

This has the same motivation as the User based search in d9c22d1.

Test Plan: unit test updated